Released

+++ ScraBoy/Features/CMS/User/UserController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNet.Identity;
 using ScraBoy.Features.CMS.Role;
+using ScraBoy.Features.CMS.Upload;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,8 +13,10 @@ namespace ScraBoy.Features.CMS.User
     
     [RoutePrefix("user")]
     [Authorize]
-    public class UserController : Controller
+    public class UserController : UploadController
     {
+        private readonly string pathFolder = "~/Image/profile/";
+        private readonly string defaultProfile = "~/Image/profile/default.jpg";
         private readonly IUserRepository usersRepository;
         private readonly IRoleRepository rolesRepository;
 
@@ -67,6 +70,7 @@ namespace ScraBoy.Features.CMS.User
         [Authorize(Roles = "admin")]
         public async Task<ActionResult> Create(UserViewModel model)
         {
+            model.UrlImage = defaultProfile;
             var completed = await userservice.CreateAsync(model);
 
             model.LoadUserRoles(await rolesRepository.GetAllRolesAsync());
@@ -75,7 +79,6 @@ namespace ScraBoy.Features.CMS.User
             {
                 return RedirectToAction("Index");
             }
-
             return View(model);
         }
         [Route("profile/{userId}")]
@@ -122,7 +125,19 @@ namespace ScraBoy.Features.CMS.User
         [ValidateAntiForgeryToken]
         [Authorize(Roles = "admin, editor, author")]
         public async Task<ActionResult> Edit(UserViewModel model, string username)
-        { 
+        {
+            var user = await userservice.GetUserByNameAsync(username);
+
+            if(user == null)
+            {
+                return HttpNotFound();
+            }
+
+            if(!ModelState.IsValid)
+            {
+                return View(model);
+            }
+      
             var currentUser = User.Identity.Name;
 
             var isAdmin = User.IsInRole("admin");
@@ -131,6 +146,31 @@ namespace ScraBoy.Features.CMS.User
             {
                 return new HttpUnauthorizedResult();
             }
+
+            if(model.ImageFile != null)
+            {
+                if(!user.UrlImage.Equals(defaultProfile))
+                {
+                    DeleteOldImage(user.UrlImage);
+                }
+                
+                var filePath = GetFullFile(model.ImageFile.FileName);
+
+                if(!CheckFileType(filePath))
+                {
+                    ModelState.AddModelError(string.Empty,"Upload Image with JPEG, JPG OR PNG Extension");
+                    return View(model);
+                }
+
+                SaveImage(model.ImageFile,pathFolder,filePath);
+
+                model.UrlImage = pathFolder + filePath;
+            }
+            else
+            {
+                model.UrlImage = user.UrlImage;
+            }
+
             var userUpdated = await this.userservice.UpdateProfile(model);
 
             if(userUpdated)
@@ -140,6 +180,7 @@ namespace ScraBoy.Features.CMS.User
                     return RedirectToAction("Index");
                 }
                 return RedirectToAction("index","admin");
+                
             }
 
             return View(model);
@@ -151,13 +192,20 @@ namespace ScraBoy.Features.CMS.User
         [Authorize(Roles = "admin")]
         public async Task<ActionResult> Delete(string username)
         {
+            var user = await userservice.GetUserByNameAsync(username);
+
+            if(user == null)
+            {
+                return HttpNotFound();
+            }
+
             if(username.Equals("Ainul"))
             {
                 ModelState.AddModelError(string.Empty,"Cannot Delete Your Account");
                 return View("Index");
             }
             await this.userservice.DeleteAsync(username);
-
+            DeleteOldImage(user.UrlImage);
             return RedirectToAction("Index");
         }
